Combine actions for deletion of file source, git source
