--- tests/Application/FileSource/AbstractDeleteTest.php
@@ -1,60 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Tests\Application\FileSource;
-
-use App\Tests\Application\AbstractApplicationTestCase;
-use App\Tests\Application\CreateSourceTrait;
-use App\Tests\Application\UnauthorizedUserDataProviderTrait;
-use SmartAssert\TestAuthenticationProviderBundle\ApiKeyProvider;
-use SmartAssert\TestAuthenticationProviderBundle\UserProvider;
-use Symfony\Component\Uid\Ulid;
-
-abstract class AbstractDeleteTest extends AbstractApplicationTestCase
-{
-    use UnauthorizedUserDataProviderTrait;
-    use AssertFileSourceTrait;
-    use CreateSourceTrait;
-
-    /**
-     * @dataProvider unauthorizedUserDataProvider
-     */
-    public function testDeleteUnauthorizedUser(?string $token): void
-    {
-        $response = $this->applicationClient->makeDeleteFileSourceRequest($token, (string) new Ulid());
-
-        self::assertSame(401, $response->getStatusCode());
-    }
-
-    public function testDeleteNotFound(): void
-    {
-        $apiKeyProvider = self::getContainer()->get(ApiKeyProvider::class);
-        \assert($apiKeyProvider instanceof ApiKeyProvider);
-        $apiKey = $apiKeyProvider->get('user@example.com');
-
-        $response = $this->applicationClient->makeDeleteFileSourceRequest($apiKey->key, (string) new Ulid());
-
-        self::assertSame(404, $response->getStatusCode());
-    }
-
-    public function testDeleteSuccess(): void
-    {
-        $apiKeyProvider = self::getContainer()->get(ApiKeyProvider::class);
-        \assert($apiKeyProvider instanceof ApiKeyProvider);
-        $apiKey = $apiKeyProvider->get('user@example.com');
-
-        $userProvider = self::getContainer()->get(UserProvider::class);
-        \assert($userProvider instanceof UserProvider);
-        $user = $userProvider->get('user@example.com');
-
-        $label = md5((string) rand());
-        $id = $this->createFileSource($apiKey->key, $label);
-
-        $getResponse = $this->applicationClient->makeGetSourceRequest($apiKey->key, $id);
-        $this->assertRetrievedFileSource($getResponse, $label, $user->id, $id);
-
-        $deleteResponse = $this->applicationClient->makeDeleteFileSourceRequest($apiKey->key, $id);
-        $this->assertDeletedFileSource($deleteResponse, $label, $user->id, $id);
-    }
-}
