+++ tests/Application/GitSource/AssertGitSourceTrait.php
@@ -49,6 +49,7 @@ trait AssertGitSourceTrait
         string $expectedHostUrl,
         string $expectedPath,
         bool $expectedHasCredentials,
+        ?string $expectedUserId = null,
     ): void {
         Assert::assertSame(200, $response->getStatusCode());
         Assert::assertSame('application/json', $response->getHeaderLine('content-type'));
@@ -59,17 +60,20 @@ trait AssertGitSourceTrait
         $deletedAt = $responseData['deleted_at'] ?? null;
         Assert::assertIsInt($deletedAt);
 
-        Assert::assertSame(
-            [
-                'id' => $expectedId,
-                'label' => $expectedLabel,
-                'type' => 'git',
-                'deleted_at' => $deletedAt,
-                'host_url' => $expectedHostUrl,
-                'path' => $expectedPath,
-                'has_credentials' => $expectedHasCredentials,
-            ],
-            $responseData
-        );
+        $expectedResponseData = [
+            'id' => $expectedId,
+            'label' => $expectedLabel,
+            'type' => 'git',
+            'host_url' => $expectedHostUrl,
+            'path' => $expectedPath,
+            'has_credentials' => $expectedHasCredentials,
+            'deleted_at' => $deletedAt,
+        ];
+
+        if (is_string($expectedUserId)) {
+            $expectedResponseData['user_id'] = $expectedUserId;
+        }
+
+        Assert::assertEquals($expectedResponseData, $responseData);
     }
 }
