--- tests/Functional/Application/GitSource/DeleteTest.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Tests\Functional\Application\GitSource;
-
-use App\Tests\Application\GitSource\AbstractDeleteTest;
-use App\Tests\Functional\GetClientAdapterTrait;
-
-class DeleteTest extends AbstractDeleteTest
-{
-    use GetClientAdapterTrait;
-}
