+++ tests/Functional/Application/Source/DeleteTest.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\Tests\Functional\Application\FileSource;
+namespace App\Tests\Functional\Application\Source;
 
-use App\Tests\Application\FileSource\AbstractDeleteTest;
+use App\Tests\Application\Source\AbstractDeleteTest;
 use App\Tests\Functional\GetClientAdapterTrait;
 
 class DeleteTest extends AbstractDeleteTest
