+++ tests/Functional/Controller/Source/GitSourceControllerTest.php
@@ -113,42 +113,6 @@ class GitSourceControllerTest extends AbstractApplicationTestCase
     }
 
     /**
-     * @dataProvider usersClientExceptionDataProvider
-     *
-     * @param array<mixed> $expectedData
-     */
-    public function testDeleteHandlesException(
-        \Exception $exception,
-        int $expectedStatusCode,
-        array $expectedData
-    ): void {
-        $apiKey = md5((string) rand());
-        $apiToken = md5((string) rand());
-        $sourceId = (string) new Ulid();
-
-        $usersClient = \Mockery::mock(UsersClient::class);
-        $usersClient
-            ->shouldReceive('createApiToken')
-            ->with($apiKey)
-            ->andReturn(new Token($apiToken))
-        ;
-
-        $gitSourceClient = \Mockery::mock(GitSourceClientInterface::class);
-        $gitSourceClient
-            ->shouldReceive('delete')
-            ->with($apiToken, $sourceId)
-            ->andThrow($exception)
-        ;
-
-        self::getContainer()->set(UsersClient::class, $usersClient);
-        self::getContainer()->set(GitSourceClientInterface::class, $gitSourceClient);
-
-        $response = $this->applicationClient->makeDeleteGitSourceRequest($apiKey, $sourceId);
-
-        $this->assertJsonResponse($response, $expectedStatusCode, $expectedData);
-    }
-
-    /**
      * @return array<mixed>
      */
     public function usersClientExceptionDataProvider(): array
