--- tests/Integration/FileSource/DeleteTest.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Tests\Integration\FileSource;
-
-use App\Tests\Application\FileSource\AbstractDeleteTest;
-use App\Tests\Integration\GetClientAdapterTrait;
-
-class DeleteTest extends AbstractDeleteTest
-{
-    use GetClientAdapterTrait;
-}
